split long file download spec into separate specs
