+++ examples/testing-dom__download/README.md
@@ -2,8 +2,6 @@
 
 ![File download in Chrome](images/chrome.png)
 
-See [cypress/plugins/index.js](cypress/plugins/index.js) to see how we validate the files after download. See the [cypress/integration/spec.js](cypress/integration/spec.js) spec file that downloads and verifies:
-
 - a CSV file
 - an Excel file
 - a PNG image
@@ -11,7 +9,16 @@ See [cypress/plugins/index.js](cypress/plugins/index.js) to see how we validate 
 - a JS file
 - a Zip file
 
-Text files are validated right from the browser spec, but the binary files like the downloaded Zip archives need to be validated from Node code using the plugins file.
+Spec file | Description
+---|---
+[form-submission-spec.js](./cypress/integration/form-submission-spec.js) | Intercepts and verifies a file downloaded after a form submission
+[local-download-spec.js](./cypress/integration/local-download-spec.js) | Downloads files from local domain by using `<a href=... download>` anchor links
+[location-href-spec.js](./cypress/integration/location-href-spec.js) | Intercepts and verifies a file downloaded via setting `document.location.href=...` URL
+[remote-download-spec.js](./cypress/integration/remote-download-spec.js) | Downloads files from another domain by using `<a href=... download>` anchor links
+
+## Notes
+
+Text files are validated right from the browser spec, but the binary files like the downloaded Zip archives need to be validated from Node code using the plugins file, see [cypress/plugins/index.js](cypress/plugins/index.js)
 
 The spec also shows how to "catch" form submission that downloads a file using [cy.intercept()](https://on.cypress.io/intercept). After intercepting the request and redirecting back at the test, the test shows how to request the file and validate it.
